+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
+## [1.11.13] - unreleased
+### Added
+### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+
 ## [1.11.12] - 6/27/26
 ### Added
 - GHORG_FETCH_GIT_LFS to run git lfs fetch --all on every repo which will correctly fetch all LFS objects from the upstream repository; thanks @dotboris